Document public RPC APIs
